fix: corrigindo as funcoes de email afim de reaproveitamento
