+++ .gitignore
@@ -4,3 +4,7 @@ yarn.lock
 .env
 yarn-error.log
 package-lock.json
+.env*
+.flaskenv*
+!.env.project
+!.env.vault
