+++ src/services/sessions/verifyLink.service.ts
@@ -4,23 +4,19 @@ import { AppError } from "../../errors/AppError";
 import jwt from "jsonwebtoken";
 import { status_code } from "../../utils/status_code";
 import bcrypt from "bcryptjs";
-import * as nodemailer from "nodemailer";
-import Mailgen from "mailgen";
+import { createEmail, sendEmail } from "../email/email.service";
 
 export const verifyLinkService = async (
   id: string,
   token: string,
-  passord: string
+  password: string
 ) => {
   const userRepository = AppDataSource.getRepository(User);
 
   const user = await userRepository.findOneBy({ id });
 
   if (!user) {
-    throw new AppError(
-      status_code.HTTP_403_FORBIDDEN,
-      "Email or password invalid"
-    );
+    throw new AppError(status_code.HTTP_404_NOT_FOUND, "User not faund");
   }
   if (!user.isActive) {
     throw new AppError(
@@ -36,22 +32,9 @@ export const verifyLinkService = async (
     }
   });
 
-  const passwordHashed = await bcrypt.hash(passord, 10);
+  const passwordHashed = await bcrypt.hash(password, 10);
   userRepository.update(id, { password: passwordHashed });
-  const email = createEmail(user, passwordHashed);
-  const emailSended = sendEmail(user.email, email);
-  return emailSended;
-};
-
-const createEmail = (user: any, password: string) => {
-  var mailGenerator = new Mailgen({
-    theme: "default",
-    product: {
-      name: "So veiculos T11 ",
-      link: `http://localhost:${process.env.PORT}/`,
-    },
-  });
-  var email = {
+  var ToSend = {
     body: {
       name: user.name,
       intro:
@@ -60,37 +43,7 @@ const createEmail = (user: any, password: string) => {
       outro: "Don't share this password with anyone",
     },
   };
-
-  // Generate an HTML email with the provided contents
-  var html = mailGenerator.generate(email);
-
-  // Generate the plaintext version of the e-mail (for clients that do not support HTML)
-  var text = mailGenerator.generatePlaintext(email);
-
-  return { html, text };
-};
-const sendEmail = (email: string, emailToSend: any) => {
-  const transporter = nodemailer.createTransport({
-    service: "gmail",
-    auth: {
-      user: process.env.NODEMAIL_USER,
-      pass: process.env.NODEMAIL_PASSWORD,
-    },
-  });
-
-  const mailOptions = {
-    from: process.env.NODEMAIL_USER,
-    to: email,
-    subject: "Password Reset",
-    html: emailToSend.html,
-    text: emailToSend.text,
-  };
-
-  transporter.sendMail(mailOptions, function (error: any, info: any) {
-    if (error) {
-      throw new AppError(status_code.HTTP_401_UNAUTHORIZED, error);
-    } else {
-      return "Email sent: " + info.response;
-    }
-  });
+  const email = createEmail(ToSend);
+  const emailSended = sendEmail(user.email, email, "Password Reseted");
+  return emailSended;
 };
